refactor: moved constants folder to types folder
